Consolidate all time functions in the same place.
