+++ package.json
@@ -2,7 +2,7 @@
   "name": "morpheus-frontend",
   "version": "0.7.0",
   "private": true,
-  "proxy": "http://localhost:8888",
+  "proxy": "https://localhost:8888",
   "dependencies": {
     "@fortawesome/fontawesome-svg-core": "1.2.30",
     "@fortawesome/free-brands-svg-icons": "5.14.0",
